refactor(kirra-planner): extract GeometricPlanner into geometric.rs (de-monolith)
lib.rs was the largest production file in the workspace (4562 lines) and a
god-root: a 9-module hub that ALSO carried the entire GeometricPlanner doer
inline (the `impl Planner for GeometricPlanner` alone was ~750 lines, plus
GeometricPlannerConfig, the corridor/geometry helpers, and LateralBump). Every
other planner (learned, learned_v2, mick) is already its own module —
GeometricPlanner was the one big one left in the root.

Move it VERBATIM into crates/kirra-planner/src/geometric.rs: LimitKind, the
STOP_EPSILON_MPS / VELOCITY_PROFILE_DS constants, GeometricPlannerConfig,
GeometricPlanner (both impls), all the private geometry helpers, LateralBump,
and the planner's ~2500-line test module (which references several of those
private helpers, so it moves with them). lib.rs re-exports
`GeometricPlanner`/`GeometricPlannerConfig`, so every external path is stable.

Behaviour is unchanged — a pure relocation:
- lib.rs 4562 -> 396 lines (now just the type vocabulary + module hub).
- geometric.rs holds the doer + its tests.
- No public API change; all 8 reverse-dependency crates build unchanged.
- kirra-planner: 195 lib tests + integration suites pass, clippy clean, fmt clean.

Dropped the now-dead `use kirra_core::containment::MAX_TRAJECTORY_HORIZON` from
lib.rs (its only user moved to geometric.rs, which imports it directly).

Part of the ongoing de-monolith program; the safety checker
(kirra_trajectory::validate_trajectory_slow) is untouched.
